- added a js confirmation before deletign a theme - fix link to css - simplified language loading (a couple of years ago during utf8 migration I left the possibility to have languages defined in another charsets - this is useless)
git-svn-id: http://piwigo.org/svn/trunk@5200 68402e56-0260-453c-a942-63ccdbb3a9ee

+++ admin/themes/default/template/themes_installed.tpl
@@ -31,7 +31,7 @@
     <div class="themeShot"><img src="{$theme.screenshot}"></div>
     <div class="themeActions">
       <a href="{$activate_baseurl}{$theme.id}" title="{'Make this theme available to users'|@translate}">{'Activate'|@translate}</a>
-      | <a href="{$delete_baseurl}{$theme.id}" title="{'Delete this theme'|@translate}">{'Delete'|@translate}</a>
+      | <a href="{$delete_baseurl}{$theme.id}" onclick="return confirm('{'Are you sure?'|@translate|@escape:javascript}');" title="{'Delete this theme'|@translate}">{'Delete'|@translate}</a>
     </div>
     
   </div>
